requirements.txt
python-telegram-bot==20.7
chromadb==0.4.22
python-dotenv==1.0.0
python-docx==1.1.0
httpx==0.25.2
sentence-transformers==5.1.1
numpy==1.26.4
fastapi==0.104.1
uvicorn==0.24.0.post1
mysql-connector-python==8.2.0
email-validator==2.1.0.post1
python-multipart==0.0.9
passlib[argon2]
